Convert away from Models.
