--- src/models/index.ts
@@ -1,5 +1,0 @@
-export { default as ArgumentsModel } from './ArgumentsModel'
-export { default as DicePoolsModel } from './DicePoolsModel'
-export { default as NotationModel } from './NotationModel'
-export { default as OptionsModel } from './OptionsModel'
-export { default as ParametersModel } from './ParametersModel'
