+++ src/utils/optionsToDescription.ts
@@ -1,0 +1,12 @@
+import type { RollOptions } from '~types'
+import {
+  formatCoreDescriptions,
+  formatModifierDescriptions
+} from './descriptionGenerators'
+
+export function optionsToDescription(options: RollOptions<number | string>) {
+  return [
+    formatCoreDescriptions(options),
+    ...formatModifierDescriptions(options)
+  ]
+}
