+++ src/validateNotation.ts
@@ -1,0 +1,24 @@
+import { isCustomSidesStringArg, isDiceNotationArg } from '~guards'
+import type { NotationValidationResult } from '~types'
+import { notationToOptions } from '~utils/notationToOptions'
+import { optionsToDescription } from './utils/optionsToDescription'
+import { optionsToNotation } from './utils/optionsToNotation'
+
+export function validateNotation(notation: string): NotationValidationResult {
+  if (!isDiceNotationArg(notation)) {
+    return {
+      valid: false,
+      description: []
+    }
+  }
+
+  const digested = notationToOptions(notation)
+
+  return {
+    valid: true,
+    digested,
+    notation: optionsToNotation(digested),
+    type: isCustomSidesStringArg(digested.sides) ? 'custom' : 'numerical',
+    description: optionsToDescription(digested)
+  }
+}
